requirements.txt
# Core
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0

# Bybit API
pybit==5.14.0
requests==2.31.0
websocket-client==1.9.0
tenacity==8.2.3

# Backtest
tqdm==4.65.0
pyarrow==12.0.0
joblib==1.3.2

# Notifications
python-telegram-bot==20.7

# Monitoring
psutil==5.9.5
